[IMP] project_utilities: tidy up code
