+++ project_utilities/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     "name": "Project Utilities",
-    "version": "11.0.2.1.0",
+    "version": "11.0.2.0.0",
     "license": "AGPL-3",
     "depends": [
         "project",
