+++ project_utilities/models/__init__.py
@@ -3,4 +3,3 @@
 
 from . import project_project
 from . import res_config_settings
-from . import project_task
