+++ project_utilities/models/project_project.py
@@ -2,8 +2,12 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from odoo.tools.safe_eval import safe_eval
-from odoo import models
+from odoo import api, fields, models
 from odoo.osv import expression
+
+from dateutil.relativedelta import relativedelta
+
+str2datetime = fields.Datetime.from_string
 
 
 class ProjectProject(models.Model):
@@ -17,3 +21,21 @@ class ProjectProject(models.Model):
         action_dict['domain'] = expression.AND(
             [new_domain, safe_eval(action_dict.get('domain') or '[]')])
         return action_dict
+
+
+class ProjectTask(models.Model):
+    _inherit = 'project.task'
+
+    @api.multi
+    def _change_project_task_date(self, new_days):
+        if not new_days:
+            return
+        for task in self:
+            vals = {}
+            deltadays = relativedelta(days=new_days)
+            if task.date_start:
+                vals['date_start'] = str2datetime(task.date_start) + deltadays
+            if task.date_end:
+                vals['date_end'] = str2datetime(task.date_end) + deltadays
+            if vals:
+                task.write(vals)
